+++ Rhygine/InputLayout.cpp
@@ -3,7 +3,7 @@
 
 InputLayout::InputLayout(std::vector< D3D11_INPUT_ELEMENT_DESC> desc, ID3DBlob* blob)
 {
-	THROW_IF_FAILED( GetDevice()->CreateInputLayout(desc.data(), desc.size(),
+	THROW_IF_FAILED( GetDevice()->CreateInputLayout(desc.data(), static_cast<UINT>(desc.size()),
 			blob->GetBufferPointer(),
 			blob->GetBufferSize(),
 			&layoutPointer)
