+++ Rhygine/TextureGenerator.h
@@ -14,37 +14,37 @@ public:
 	/// </summary>
 	/// <param name="width">The width of the texture.</param>
 	/// <param name="height">The height of the texture.</param>
-	TextureGenerator(size_t width, size_t height);
+	TextureGenerator(int width, int height);
 	/// <summary>
 	/// Color constructor.
 	/// </summary>
 	/// <param name="width">The width of the texture.</param>
 	/// <param name="height">The height of the texture.</param>
 	/// <param name="initialColor">The initial color that each pixel of the texture has.</param>
-	TextureGenerator(size_t width, size_t height, RhyC::color initialColor);
+	TextureGenerator(int width, int height, RhyC::color initialColor);
 	/// <summary>
 	/// Exchanges the color of all touching pixels that share the same color as the specified pixel with the
 	/// brush color.
 	/// </summary>
 	/// <param name="x">The x coord of the pixel.</param>
 	/// <param name="y">The y coord of the pixel.</param>
-	void Fill(size_t x, size_t y);
+	void Fill(int x, int y);
 	//void DrawLine(int xFrom, int yFrom, int xTo, int yTo);
 	/// <summary>
 	/// Replaces the color of a single pixel with the color of the brush.
 	/// </summary>
 	/// <param name="x">The x coord of the pixel.</param>
 	/// <param name="y">The y coord of the pixel.</param>
-	void DrawPixel(size_t x, size_t y);
+	void DrawPixel(int x, int y);
 
 	/// <summary>
 	/// Gets the width.
 	/// </summary>
-	size_t GetWidth();
+	int GetWidth();
 	/// <summary>
 	/// Gets the height.
 	/// </summary>
-	size_t GetHeight();
+	int GetHeight();
 
 	/// <summary>
 	/// Gets the color array.
@@ -58,5 +58,5 @@ public:
 	RhyC::color brush = RhyC::white;
 private:
 	std::vector<RhyC::color> tex;
-	size_t width, height;
+	int width, height;
 };
